+++ src/aiologic/meta/__init__.py
@@ -46,3 +46,7 @@ from ._markers import (
 from ._modules import (
     resolve_name as resolve_name,
 )
+from ._types import (
+    coroutine as coroutine,
+    generator as generator,
+)